another_version/requirements.txt
PyYAML
ansible-core
Jinja2